week 7

+++ smt1/dasproprak/week7/jobsheet/WhileKelipatan.java
@@ -16,7 +16,7 @@ public class WhileKelipatan {
                 counter++;
             }
         }
-        float avg = jumlah / counter; 
+        float avg = (float) jumlah / counter ;
         System.out.printf("Banyaknya bilangan %d dari 1 sampai 50 adalah %d\n", kelipatan, counter);
         System.out.printf("Total bilangan kelipatan %d dari 1 sampai 50 adalah %d\n", kelipatan, jumlah);
         System.out.printf("Rata rata dari bilangan kelipatan %d adalah %.2f\n " , kelipatan, avg);
